improve docs for MakePublicKey to state mutation of privateKey
MakePublicKey takes in a privateKey and mutates it setting
the first 32 bytes to those original 32 bytes used to
generate the publicKey.
This wasn't made clear in the comments and required digging
into the actual implementation when I was advocating for
https://github.com/tendermint/go-crypto/issues/31.
I noticed that in crypto
https://github.com/tendermint/go-crypto/blob/1bc8de4caa844f8b64c120e65b047898f22b7f3e/priv_key.go#L99-L114

there was an unexplained call to ed25519.MakePublicKey nor were
there any comments to explain that.